+++ webpack.config.js
@@ -6,6 +6,8 @@ const HtmlWebpackPlugin = require("html-webpack-plugin");
 const BundleAnalyzerPlugin = require("webpack-bundle-analyzer").BundleAnalyzerPlugin;
 const UglifyJsPlugin = require("uglifyjs-webpack-plugin");
 
+const pkg = require("./package.json");
+
 module.exports = {
   entry: {
     build: "./src/App.tsx",
@@ -132,6 +134,7 @@ if (process.env.NODE_ENV === "production") {
       "process.env": {
         NODE_ENV: "\"production\"",
       },
+      "VERSION": `"${pkg.version}"`,
     }),
     new webpack.LoaderOptionsPlugin({
       debug: false,
@@ -149,6 +152,9 @@ if (process.env.NODE_ENV === "production") {
 } else {
   module.exports.plugins = (module.exports.plugins || []).concat([
     // new BundleAnalyzerPlugin(),
+    new webpack.DefinePlugin({
+      "VERSION": `"${pkg.version}"`,
+    }),
   ]);
   // https://github.com/webpack/webpack/issues/6642#issuecomment-370222543
   module.exports.output.globalObject = 'this';