+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## NEXT
+* Added version indicator on header / in about.
+
 ## 1.5.4
 * Centered new loading animation.
 * Cleaned up CopyAndEditPaste. Fixed bug where, when submitting, if there were an error the edited paste would be lost.